third_party/blink/web_tests/external/wpt/encoding/legacy-mb-schinese/gb18030/gb18030-decoder.any.worker-expected.txt

This is a testharness.js-based test.
Found 238 tests; 13 PASS, 225 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS gb18030 decoder: ASCII
FAIL gb18030 decoder: euro assert_equals: expected "€" but got "\ufffd"
FAIL gb18030 decoder: initial byte out of accepted ranges assert_equals: expected "\ufffd" but got ""
PASS gb18030 decoder: end of queue, gb18030 first not 0
PASS gb18030 decoder: two bytes 0x81 0x28
PASS gb18030 decoder: two bytes 0x81 0x40
PASS gb18030 decoder: two bytes 0x81 0x7e
PASS gb18030 decoder: two bytes 0x81 0x7f
PASS gb18030 decoder: two bytes 0x81 0x80
PASS gb18030 decoder: two bytes 0x81 0xFE
FAIL gb18030 decoder: two bytes 0x81 0xFF assert_equals: expected "\ufffd" but got "\ufffd"
PASS gb18030 decoder: two bytes 0xFE 0x40
PASS gb18030 decoder: two bytes 0xFE 0xFE
FAIL gb18030 decoder: two bytes 0xFE 0xFF assert_equals: expected "\ufffd" but got "\ufffd"
FAIL gb18030 decoder: two bytes 0x81 0x30 assert_equals: expected "\ufffd" but got "\ufffd0"
FAIL gb18030 decoder: three bytes 0x81 0x30 0xFE assert_equals: expected "\ufffd" but got "\ufffd0\ufffd"
FAIL gb18030 decoder: three bytes 0x81 0x30 0xFF assert_equals: expected "\ufffd0\ufffd" but got "\ufffd0"
PASS gb18030 decoder: four bytes 0x81 0x30 0xFE 0x29
FAIL gb18030 decoder: four bytes 0xFE 0x39 0xFE 0x39 assert_equals: expected "\ufffd" but got "\ufffd9\ufffd9"
FAIL gb18030 decoder: pointer 7458 assert_equals: expected "Ḿ" but got "\ufffd5\ufffd6"
FAIL gb18030 decoder: pointer 7457 assert_equals: expected "" but got "\ufffd5\ufffd7"
FAIL gb18030 decoder: pointer 7459 assert_equals: expected "Ṁ" but got "\ufffd5\ufffd8"
FAIL gb18030 decoder: pointer 39419 assert_equals: expected "\uffff" but got "\ufffd1\ufffd9"
FAIL gb18030 decoder: pointer 39420 assert_equals: expected "\ufffd" but got "\ufffd1\ufffd0"
FAIL gb18030 decoder: pointer 189999 assert_equals: expected "\ufffd" but got "\ufffd9\ufffd9"
FAIL gb18030 decoder: pointer 189000 assert_equals: expected "𐀀" but got "\ufffd0\ufffd0"
FAIL gb18030 decoder: pointer 1237575 assert_equals: expected "􏿿" but got "\ufffd2\ufffd5"
FAIL gb18030 decoder: pointer 1237576 assert_equals: expected "\ufffd" but got "\ufffd2\ufffd6"
FAIL gb18030 decoder: legacy ICU special case 1 assert_equals: expected "" but got "\ufffd6\ufffd0"
PASS gb18030 decoder: legacy ICU special case 2
PASS gb18030 decoder: legacy ICU special case 3
FAIL gb18030 decoder: range 0 assert_equals: expected "" but got "\ufffd0\ufffd0"
FAIL gb18030 decoder: range 1 assert_equals: expected "¥" but got "\ufffd0\ufffd6"
FAIL gb18030 decoder: range 2 assert_equals: expected "©" but got "\ufffd0\ufffd8"
FAIL gb18030 decoder: range 3 assert_equals: expected "²" but got "\ufffd0\ufffd5"
FAIL gb18030 decoder: range 4 assert_equals: expected "¸" but got "\ufffd0\ufffd0"
FAIL gb18030 decoder: range 5 assert_equals: expected "Ø" but got "\ufffd0\ufffd1"
FAIL gb18030 decoder: range 6 assert_equals: expected "â" but got "\ufffd0\ufffd9"
FAIL gb18030 decoder: range 7 assert_equals: expected "ë" but got "\ufffd0\ufffd5"
FAIL gb18030 decoder: range 8 assert_equals: expected "î" but got "\ufffd0\ufffd6"
FAIL gb18030 decoder: range 9 assert_equals: expected "ô" but got "\ufffd0\ufffd0"
FAIL gb18030 decoder: range 10 assert_equals: expected "ø" but got "\ufffd0\ufffd3"
FAIL gb18030 decoder: range 11 assert_equals: expected "û" but got "\ufffd0\ufffd4"
FAIL gb18030 decoder: range 12 assert_equals: expected "ý" but got "\ufffd0\ufffd5"
FAIL gb18030 decoder: range 13 assert_equals: expected "Ă" but got "\ufffd0\ufffd9"
FAIL gb18030 decoder: range 14 assert_equals: expected "Ĕ" but got "\ufffd0\ufffd6"
FAIL gb18030 decoder: range 15 assert_equals: expected "Ĝ" but got "\ufffd0\ufffd3"
FAIL gb18030 decoder: range 16 assert_equals: expected "Ĭ" but got "\ufffd0\ufffd8"
FAIL gb18030 decoder: range 17 assert_equals: expected "Ņ" but got "\ufffd0\ufffd2"
FAIL gb18030 decoder: range 18 assert_equals: expected "ŉ" but got "\ufffd0\ufffd5"
FAIL gb18030 decoder: range 19 assert_equals: expected "Ŏ" but got "\ufffd0\ufffd9"
FAIL gb18030 decoder: range 20 assert_equals: expected "Ŭ" but got "\ufffd0\ufffd8"
FAIL gb18030 decoder: range 21 assert_equals: expected "Ǐ" but got "\ufffd0\ufffd6"
FAIL gb18030 decoder: range 22 assert_equals: expected "Ǒ" but got "\ufffd0\ufffd7"
FAIL gb18030 decoder: range 23 assert_equals: expected "Ǔ" but got "\ufffd0\ufffd8"
FAIL gb18030 decoder: range 24 assert_equals: expected "Ǖ" but got "\ufffd0\ufffd9"
FAIL gb18030 decoder: range 25 assert_equals: expected "Ǘ" but got "\ufffd0\ufffd0"
FAIL gb18030 decoder: range 26 assert_equals: expected "Ǚ" but got "\ufffd0\ufffd1"
FAIL gb18030 decoder: range 27 assert_equals: expected "Ǜ" but got "\ufffd0\ufffd2"
FAIL gb18030 decoder: range 28 assert_equals: expected "ǝ" but got "\ufffd0\ufffd3"
FAIL gb18030 decoder: range 29 assert_equals: expected "Ǻ" but got "\ufffd0\ufffd1"
FAIL gb18030 decoder: range 30 assert_equals: expected "ɒ" but got "\ufffd0\ufffd8"
FAIL gb18030 decoder: range 31 assert_equals: expected "ɢ" but got "\ufffd0\ufffd3"
FAIL gb18030 decoder: range 32 assert_equals: expected "ˈ" but got "\ufffd0\ufffd4"
FAIL gb18030 decoder: range 33 assert_equals: expected "ˌ" but got "\ufffd0\ufffd5"
FAIL gb18030 decoder: range 34 assert_equals: expected "˚" but got "\ufffd0\ufffd8"
FAIL gb18030 decoder: range 35 assert_equals: expected "΢" but got "\ufffd0\ufffd1"
FAIL gb18030 decoder: range 36 assert_equals: expected "Ϊ" but got "\ufffd0\ufffd2"
FAIL gb18030 decoder: range 37 assert_equals: expected "ς" but got "\ufffd0\ufffd9"
FAIL gb18030 decoder: range 38 assert_equals: expected "ϊ" but got "\ufffd0\ufffd0"
FAIL gb18030 decoder: range 39 assert_equals: expected "Ђ" but got "\ufffd0\ufffd5"
FAIL gb18030 decoder: range 40 assert_equals: expected "ѐ" but got "\ufffd0\ufffd9"
FAIL gb18030 decoder: range 41 assert_equals: expected "ђ" but got "\ufffd0\ufffd0"
FAIL gb18030 decoder: range 42 assert_equals: expected "‑" but got "\ufffd6\ufffd2"
FAIL gb18030 decoder: range 43 assert_equals: expected "‗" but got "\ufffd6\ufffd4"
FAIL gb18030 decoder: range 44 assert_equals: expected "‚" but got "\ufffd6\ufffd5"
FAIL gb18030 decoder: range 45 assert_equals: expected "„" but got "\ufffd6\ufffd7"
FAIL gb18030 decoder: range 46 assert_equals: expected "‧" but got "\ufffd6\ufffd4"
FAIL gb18030 decoder: range 47 assert_equals: expected "‱" but got "\ufffd6\ufffd3"
FAIL gb18030 decoder: range 48 assert_equals: expected "‴" but got "\ufffd6\ufffd4"
FAIL gb18030 decoder: range 49 assert_equals: expected "‶" but got "\ufffd6\ufffd5"
FAIL gb18030 decoder: range 50 assert_equals: expected "‼" but got "\ufffd6\ufffd0"
FAIL gb18030 decoder: range 51 assert_equals: expected "₭" but got "\ufffd6\ufffd2"
FAIL gb18030 decoder: range 52 assert_equals: expected "℄" but got "\ufffd6\ufffd8"
FAIL gb18030 decoder: range 53 assert_equals: expected "℆" but got "\ufffd6\ufffd9"
FAIL gb18030 decoder: range 54 assert_equals: expected "ℊ" but got "\ufffd6\ufffd2"
FAIL gb18030 decoder: range 55 assert_equals: expected "℗" but got "\ufffd6\ufffd4"
FAIL gb18030 decoder: range 56 assert_equals: expected "™" but got "\ufffd6\ufffd4"
FAIL gb18030 decoder: range 57 assert_equals: expected "Ⅼ" but got "\ufffd6\ufffd6"
FAIL gb18030 decoder: range 58 assert_equals: expected "ⅺ" but got "\ufffd6\ufffd0"
FAIL gb18030 decoder: range 59 assert_equals: expected "↔" but got "\ufffd6\ufffd2"
FAIL gb18030 decoder: range 60 assert_equals: expected "↚" but got "\ufffd6\ufffd4"
FAIL gb18030 decoder: range 61 assert_equals: expected "∉" but got "\ufffd6\ufffd4"
FAIL gb18030 decoder: range 62 assert_equals: expected "∐" but got "\ufffd6\ufffd0"
FAIL gb18030 decoder: range 63 assert_equals: expected "−" but got "\ufffd6\ufffd1"
FAIL gb18030 decoder: range 64 assert_equals: expected "∖" but got "\ufffd6\ufffd4"
FAIL gb18030 decoder: range 65 assert_equals: expected "∛" but got "\ufffd6\ufffd8"
FAIL gb18030 decoder: range 66 assert_equals: expected "∡" but got "\ufffd6\ufffd0"
FAIL gb18030 decoder: range 67 assert_equals: expected "∤" but got "\ufffd6\ufffd2"
FAIL gb18030 decoder: range 68 assert_equals: expected "∦" but got "\ufffd6\ufffd3"
FAIL gb18030 decoder: range 69 assert_equals: expected "∬" but got "\ufffd6\ufffd4"
FAIL gb18030 decoder: range 70 assert_equals: expected "∯" but got "\ufffd6\ufffd6"
FAIL gb18030 decoder: range 71 assert_equals: expected "∸" but got "\ufffd6\ufffd1"
FAIL gb18030 decoder: range 72 assert_equals: expected "∾" but got "\ufffd6\ufffd6"
FAIL gb18030 decoder: range 73 assert_equals: expected "≉" but got "\ufffd6\ufffd6"
FAIL gb18030 decoder: range 74 assert_equals: expected "≍" but got "\ufffd6\ufffd9"
FAIL gb18030 decoder: range 75 assert_equals: expected "≓" but got "\ufffd6\ufffd4"
FAIL gb18030 decoder: range 76 assert_equals: expected "≢" but got "\ufffd6\ufffd7"
FAIL gb18030 decoder: range 77 assert_equals: expected "≨" but got "\ufffd6\ufffd9"
FAIL gb18030 decoder: range 78 assert_equals: expected "≰" but got "\ufffd6\ufffd5"
FAIL gb18030 decoder: range 79 assert_equals: expected "⊖" but got "\ufffd6\ufffd2"
FAIL gb18030 decoder: range 80 assert_equals: expected "⊚" but got "\ufffd6\ufffd5"
FAIL gb18030 decoder: range 81 assert_equals: expected "⊦" but got "\ufffd6\ufffd6"
FAIL gb18030 decoder: range 82 assert_equals: expected "⋀" but got "\ufffd6\ufffd1"
FAIL gb18030 decoder: range 83 assert_equals: expected "⌓" but got "\ufffd6\ufffd3"
FAIL gb18030 decoder: range 84 assert_equals: expected "⑪" but got "\ufffd7\ufffd6"
FAIL gb18030 decoder: range 85 assert_equals: expected "⒜" but got "\ufffd7\ufffd6"
FAIL gb18030 decoder: range 86 assert_equals: expected "╌" but got "\ufffd7\ufffd6"
FAIL gb18030 decoder: range 87 assert_equals: expected "╴" but got "\ufffd7\ufffd0"
FAIL gb18030 decoder: range 88 assert_equals: expected "▐" but got "\ufffd7\ufffd3"
FAIL gb18030 decoder: range 89 assert_equals: expected "▖" but got "\ufffd7\ufffd6"
FAIL gb18030 decoder: range 90 assert_equals: expected "▢" but got "\ufffd7\ufffd6"
FAIL gb18030 decoder: range 91 assert_equals: expected "▴" but got "\ufffd7\ufffd2"
FAIL gb18030 decoder: range 92 assert_equals: expected "▾" but got "\ufffd7\ufffd0"
FAIL gb18030 decoder: range 93 assert_equals: expected "◈" but got "\ufffd7\ufffd8"
FAIL gb18030 decoder: range 94 assert_equals: expected "◌" but got "\ufffd7\ufffd1"
FAIL gb18030 decoder: range 95 assert_equals: expected "◐" but got "\ufffd7\ufffd3"
FAIL gb18030 decoder: range 96 assert_equals: expected "◦" but got "\ufffd7\ufffd1"
FAIL gb18030 decoder: range 97 assert_equals: expected "☇" but got "\ufffd7\ufffd2"
FAIL gb18030 decoder: range 98 assert_equals: expected "☊" but got "\ufffd7\ufffd4"
FAIL gb18030 decoder: range 99 assert_equals: expected "♁" but got "\ufffd7\ufffd8"
FAIL gb18030 decoder: range 100 assert_equals: expected "♃" but got "\ufffd7\ufffd9"
FAIL gb18030 decoder: range 101 assert_equals: expected "⺂" but got "\ufffd8\ufffd9"
FAIL gb18030 decoder: range 102 assert_equals: expected "⺅" but got "\ufffd8\ufffd1"
FAIL gb18030 decoder: range 103 assert_equals: expected "⺉" but got "\ufffd8\ufffd4"
FAIL gb18030 decoder: range 104 assert_equals: expected "⺍" but got "\ufffd8\ufffd6"
FAIL gb18030 decoder: range 105 assert_equals: expected "⺘" but got "\ufffd9\ufffd6"
FAIL gb18030 decoder: range 106 assert_equals: expected "⺨" but got "\ufffd9\ufffd1"
FAIL gb18030 decoder: range 107 assert_equals: expected "⺫" but got "\ufffd9\ufffd3"
FAIL gb18030 decoder: range 108 assert_equals: expected "⺯" but got "\ufffd9\ufffd6"
FAIL gb18030 decoder: range 109 assert_equals: expected "⺴" but got "\ufffd9\ufffd0"
FAIL gb18030 decoder: range 110 assert_equals: expected "⺸" but got "\ufffd9\ufffd2"
FAIL gb18030 decoder: range 111 assert_equals: expected "⺼" but got "\ufffd9\ufffd5"
FAIL gb18030 decoder: range 112 assert_equals: expected "⻋" but got "\ufffd9\ufffd9"
FAIL gb18030 decoder: range 113 assert_equals: expected "⿼" but got "\ufffd9\ufffd2"
FAIL gb18030 decoder: range 114 assert_equals: expected "〄" but got "\ufffd9\ufffd6"
FAIL gb18030 decoder: range 115 assert_equals: expected "〘" but got "\ufffd9\ufffd7"
FAIL gb18030 decoder: range 116 assert_equals: expected "〟" but got "\ufffd9\ufffd2"
FAIL gb18030 decoder: range 117 assert_equals: expected "〪" but got "\ufffd9\ufffd4"
FAIL gb18030 decoder: range 118 assert_equals: expected "〿" but got "\ufffd9\ufffd4"
FAIL gb18030 decoder: range 119 assert_equals: expected "ゔ" but got "\ufffd9\ufffd6"
FAIL gb18030 decoder: range 120 assert_equals: expected "ゟ" but got "\ufffd9\ufffd3"
FAIL gb18030 decoder: range 121 assert_equals: expected "ヷ" but got "\ufffd9\ufffd5"
FAIL gb18030 decoder: range 122 assert_equals: expected "ヿ" but got "\ufffd9\ufffd0"
FAIL gb18030 decoder: range 123 assert_equals: expected "ㄪ" but got "\ufffd9\ufffd6"
FAIL gb18030 decoder: range 124 assert_equals: expected "㈪" but got "\ufffd9\ufffd2"
FAIL gb18030 decoder: range 125 assert_equals: expected "㈲" but got "\ufffd9\ufffd9"
FAIL gb18030 decoder: range 126 assert_equals: expected "㊤" but got "\ufffd9\ufffd2"
FAIL gb18030 decoder: range 127 assert_equals: expected "㎐" but got "\ufffd9\ufffd6"
FAIL gb18030 decoder: range 128 assert_equals: expected "㎟" but got "\ufffd9\ufffd8"
FAIL gb18030 decoder: range 129 assert_equals: expected "㎢" but got "\ufffd9\ufffd0"
FAIL gb18030 decoder: range 130 assert_equals: expected "㏅" but got "\ufffd9\ufffd4"
FAIL gb18030 decoder: range 131 assert_equals: expected "㏏" but got "\ufffd9\ufffd3"
FAIL gb18030 decoder: range 132 assert_equals: expected "㏓" but got "\ufffd9\ufffd5"
FAIL gb18030 decoder: range 133 assert_equals: expected "㏖" but got "\ufffd9\ufffd7"
FAIL gb18030 decoder: range 134 assert_equals: expected "㑈" but got "\ufffd9\ufffd0"
FAIL gb18030 decoder: range 135 assert_equals: expected "㑴" but got "\ufffd9\ufffd3"
FAIL gb18030 decoder: range 136 assert_equals: expected "㖟" but got "\ufffd0\ufffd1"
FAIL gb18030 decoder: range 137 assert_equals: expected "㘏" but got "\ufffd0\ufffd2"
FAIL gb18030 decoder: range 138 assert_equals: expected "㘛" but got "\ufffd0\ufffd3"
FAIL gb18030 decoder: range 139 assert_equals: expected "㤙" but got "\ufffd0\ufffd8"
FAIL gb18030 decoder: range 140 assert_equals: expected "㥯" but got "\ufffd0\ufffd3"
FAIL gb18030 decoder: range 141 assert_equals: expected "㧑" but got "\ufffd1\ufffd9"
FAIL gb18030 decoder: range 142 assert_equals: expected "㧠" but got "\ufffd1\ufffd3"
FAIL gb18030 decoder: range 143 assert_equals: expected "㩴" but got "\ufffd1\ufffd0"
FAIL gb18030 decoder: range 144 assert_equals: expected "㭏" but got "\ufffd1\ufffd8"
FAIL gb18030 decoder: range 145 assert_equals: expected "㱯" but got "\ufffd1\ufffd5"
FAIL gb18030 decoder: range 146 assert_equals: expected "㳡" but got "\ufffd1\ufffd8"
FAIL gb18030 decoder: range 147 assert_equals: expected "䁗" but got "\ufffd2\ufffd3"
FAIL gb18030 decoder: range 148 assert_equals: expected "䅠" but got "\ufffd2\ufffd7"
FAIL gb18030 decoder: range 149 assert_equals: expected "䌸" but got "\ufffd2\ufffd8"
FAIL gb18030 decoder: range 150 assert_equals: expected "䎭" but got "\ufffd3\ufffd4"
FAIL gb18030 decoder: range 151 assert_equals: expected "䎲" but got "\ufffd3\ufffd8"
FAIL gb18030 decoder: range 152 assert_equals: expected "䏞" but got "\ufffd3\ufffd1"
FAIL gb18030 decoder: range 153 assert_equals: expected "䓗" but got "\ufffd3\ufffd9"
FAIL gb18030 decoder: range 154 assert_equals: expected "䙍" but got "\ufffd3\ufffd2"
FAIL gb18030 decoder: range 155 assert_equals: expected "䙢" but got "\ufffd3\ufffd2"
FAIL gb18030 decoder: range 156 assert_equals: expected "䜤" but got "\ufffd3\ufffd5"
FAIL gb18030 decoder: range 157 assert_equals: expected "䜪" but got "\ufffd3\ufffd0"
FAIL gb18030 decoder: range 158 assert_equals: expected "䝽" but got "\ufffd3\ufffd2"
FAIL gb18030 decoder: range 159 assert_equals: expected "䞎" but got "\ufffd3\ufffd8"
FAIL gb18030 decoder: range 160 assert_equals: expected "䥈" but got "\ufffd4\ufffd9"
FAIL gb18030 decoder: range 161 assert_equals: expected "䥻" but got "\ufffd4\ufffd9"
FAIL gb18030 decoder: range 162 assert_equals: expected "䥾" but got "\ufffd4\ufffd1"
FAIL gb18030 decoder: range 163 assert_equals: expected "䦄" but got "\ufffd4\ufffd5"
FAIL gb18030 decoder: range 164 assert_equals: expected "䦇" but got "\ufffd4\ufffd6"
FAIL gb18030 decoder: range 165 assert_equals: expected "䦜" but got "\ufffd4\ufffd6"
FAIL gb18030 decoder: range 166 assert_equals: expected "䦠" but got "\ufffd4\ufffd9"
FAIL gb18030 decoder: range 167 assert_equals: expected "䦸" but got "\ufffd4\ufffd1"
FAIL gb18030 decoder: range 168 assert_equals: expected "䱸" but got "\ufffd4\ufffd4"
FAIL gb18030 decoder: range 169 assert_equals: expected "䲤" but got "\ufffd4\ufffd3"
FAIL gb18030 decoder: range 170 assert_equals: expected "䴚" but got "\ufffd4\ufffd4"
FAIL gb18030 decoder: range 171 assert_equals: expected "䶯" but got "\ufffd5\ufffd2"
FAIL gb18030 decoder: range 172 assert_equals: expected "龦" but got "\ufffd5\ufffd3"
FAIL gb18030 decoder: range 173 assert_equals: expected "" but got "\ufffd6\ufffd9"
FAIL gb18030 decoder: range 174 assert_equals: expected "" but got "\ufffd6\ufffd0"
FAIL gb18030 decoder: range 175 assert_equals: expected "" but got "\ufffd6\ufffd1"
FAIL gb18030 decoder: range 176 assert_equals: expected "" but got "\ufffd6\ufffd4"
FAIL gb18030 decoder: range 177 assert_equals: expected "" but got "\ufffd6\ufffd5"
FAIL gb18030 decoder: range 178 assert_equals: expected "" but got "\ufffd6\ufffd0"
FAIL gb18030 decoder: range 179 assert_equals: expected "" but got "\ufffd6\ufffd7"
FAIL gb18030 decoder: range 180 assert_equals: expected "" but got "\ufffd6\ufffd1"
FAIL gb18030 decoder: range 181 assert_equals: expected "" but got "\ufffd6\ufffd5"
FAIL gb18030 decoder: range 182 assert_equals: expected "" but got "\ufffd6\ufffd3"
FAIL gb18030 decoder: range 183 assert_equals: expected "" but got "\ufffd6\ufffd0"
FAIL gb18030 decoder: range 184 assert_equals: expected "" but got "\ufffd6\ufffd6"
FAIL gb18030 decoder: range 185 assert_equals: expected "" but got "\ufffd6\ufffd0"
FAIL gb18030 decoder: range 186 assert_equals: expected "來" but got "\ufffd0\ufffd5"
FAIL gb18030 decoder: range 187 assert_equals: expected "梁" but got "\ufffd0\ufffd1"
FAIL gb18030 decoder: range 188 assert_equals: expected "練" but got "\ufffd0\ufffd8"
FAIL gb18030 decoder: range 189 assert_equals: expected "裡" but got "\ufffd0\ufffd9"
FAIL gb18030 decoder: range 190 assert_equals: expected "鱗" but got "\ufffd0\ufffd8"
FAIL gb18030 decoder: range 191 assert_equals: expected "塚" but got "\ufffd0\ufffd4"
FAIL gb18030 decoder: range 192 assert_equals: expected "晴" but got "\ufffd0\ufffd5"
FAIL gb18030 decoder: range 193 assert_equals: expected "凞" but got "\ufffd0\ufffd6"
FAIL gb18030 decoder: range 194 assert_equals: expected "神" but got "\ufffd0\ufffd9"
FAIL gb18030 decoder: range 195 assert_equals: expected "諸" but got "\ufffd0\ufffd5"
FAIL gb18030 decoder: range 196 assert_equals: expected "逸" but got "\ufffd0\ufffd6"
FAIL gb18030 decoder: range 197 assert_equals: expected "飯" but got "\ufffd0\ufffd8"
FAIL gb18030 decoder: range 198 assert_equals: expected "︲" but got "\ufffd1\ufffd8"
FAIL gb18030 decoder: range 199 assert_equals: expected "﹅" but got "\ufffd1\ufffd9"
FAIL gb18030 decoder: range 200 assert_equals: expected "﹓" but got "\ufffd1\ufffd3"
FAIL gb18030 decoder: range 201 assert_equals: expected "﹘" but got "\ufffd1\ufffd4"
FAIL gb18030 decoder: range 202 assert_equals: expected "﹧" but got "\ufffd1\ufffd5"
FAIL gb18030 decoder: range 203 assert_equals: expected "﹬" but got "\ufffd1\ufffd6"
FAIL gb18030 decoder: range 204 assert_equals: expected "｟" but got "\ufffd1\ufffd5"
FAIL gb18030 decoder: range 205 assert_equals: expected "￦" but got "\ufffd1\ufffd4"
FAIL gb18030 decoder: range 206 assert_equals: expected "𐀀" but got "\ufffd0\ufffd0"
Harness: the test ran to completion.
